refactor openapi code & move to a class
