+++ testing_application/app.py
@@ -1,25 +1,27 @@
-from robyn import Robyn, Request, jsonify
+from robyn import Robyn, Request, jsonify, OpenAPI
 
 app = Robyn(
     file_object=__file__,
-    openapi_title="Sample Pet Store App",
-    openapi_summary=" A pet store manager.",
-    openapi_description=" This is a sample server for a pet store.",
-    openapi_terms_of_service=" https://example.com/terms/",
-    openapi_version=" 1.0.1",
-    openapi_contact_name="API Support",
-    openapi_contact_url="https://www.example.com/support",
-    openapi_contact_email="support@example.com",
-    openapi_license_name="Apache 2.0",
-    openapi_license_url="https://www.apache.org/licenses/LICENSE-2.0.html",
-    # openapi_servers=[
-    #     {"url": "/", "description": "Debug environment"},
-    #     {"url": "https://example.com/api/v1/", "description": "Production environment"},
-    # ],
-    # openapi_external_docs={
-    #     "description": "Find more info here",
-    #     "url": "https://example.com/"
-    # }
+    openapi=OpenAPI(
+        openapi_title="Sample Pet Store App",
+        openapi_summary=" A pet store manager.",
+        openapi_description=" This is a sample server for a pet store.",
+        openapi_terms_of_service=" https://example.com/terms/",
+        openapi_version=" 1.0.1",
+        openapi_contact_name="API Support",
+        openapi_contact_url="https://www.example.com/support",
+        openapi_contact_email="support@example.com",
+        openapi_license_name="Apache 2.0",
+        openapi_license_url="https://www.apache.org/licenses/LICENSE-2.0.html",
+        # openapi_servers=[
+        #     {"url": "/", "description": "Debug environment"},
+        #     {"url": "https://example.com/api/v1/", "description": "Production environment"},
+        # ],
+        # openapi_external_docs={
+        #     "description": "Find more info here",
+        #     "url": "https://example.com/"
+        # }
+    ),
 )
 
 
@@ -43,3 +45,8 @@ async def delete_user(r: Request):
 
 if __name__ == "__main__":
     app.start()
+
+# query params ->> typed dict
+# subrouter impl for openapi
+
+# add_direcotry
